Issue 3491: Fiddled a bit with indexing robustness. Build index/1 and index/2 directories instead of index and index.new.
git-svn-id: https://svn.imcode.com/imcode/imcms/trunk@4526 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ server/src/imcode/server/Config.java
@@ -17,7 +17,7 @@ public class Config {
     private String defaultLanguage;
     private String sessionCookieDomain;
     private String fileAdminRootPaths;
-    private int indexingSchedulePeriodInMinutes = 0 ;
+    private float indexingSchedulePeriodInMinutes = 0 ;
     private String documentPathPrefix;
     private int documentCacheMaxSize = 100 ;
     private String keyStorePath ;
@@ -110,11 +110,11 @@ public class Config {
         this.fileAdminRootPaths = fileAdminRootPaths;
     }
 
-    public int getIndexingSchedulePeriodInMinutes() {
+    public float getIndexingSchedulePeriodInMinutes() {
         return indexingSchedulePeriodInMinutes;
     }
 
-    public void setIndexingSchedulePeriodInMinutes( int indexingSchedulePeriodInMinutes ) {
+    public void setIndexingSchedulePeriodInMinutes( float indexingSchedulePeriodInMinutes ) {
         this.indexingSchedulePeriodInMinutes = indexingSchedulePeriodInMinutes;
     }
